delete_bq/requirements_cloud_func.txt
google-cloud-bigquery==1.10.0
pandas==0.24.2
python-dateutil==2.8.0